Balance Reminder for Dual Sim : Bug - Reminder is created whenever the balance fragment is created Deductions Fragment is now functional Encryption added for the regexes do not forget the Key and IV store in in Important data SimplyV2 folder Background sync added to regex updation, manual testing was successful need to do one programmatic testing. With debug info added it is good to release

+++ iBalance/src/main/java/com/builder/ibalance/database/helpers/DateDurationMapHelper.java
@@ -59,9 +59,9 @@ public class DateDurationMapHelper
       builder.append("Out Duration "+": "+c.getFloat(3));
       builder.append("\n");
       temp = new DateDurationModel(
-              (int)Math.ceil(c.getInt(0)),
+              (int)Math.ceil(c.getFloat(0)),
               (int)Math.ceil(c.getFloat(1)/60),
-              (int)Math.ceil(c.getInt(2)),
+              (int)Math.ceil(c.getFloat(2)),
               (int)Math.ceil(c.getFloat(3)/60),
               c.getInt(4));
       avgDateDurationDetail.add(temp);
